third_party/blink/web_tests/wpt_internal/route/parsing/route-valid.html

<!DOCTYPE html>
<title>@route rules</title>
<link rel="author" title="Morten Stenshorne" href="mailto:mstensho@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-navigation-1/#at-route">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
  test(()=> {
    const canonical = `@route --rte {\n}`;
    test_valid_rule(canonical);
    test_valid_rule(`@route  --rte{}`, canonical);
  }, "no descriptors");

  test(()=> {
    test_valid_rule("@route --ahoi\\(hoi {}", "@route --ahoi\\(hoi {\n}");
  }, "escape identifiers");
</script>
